requirements.txt
python-dotenv==1.0.1
pandas
numpy==1.26.4
scikit-learn
pymongo
certifi
pymongo[srv]==3.11
scikit-learn

# - e .